requirements.txt
Django>=3.0.6<3.1.0
djangorestframework>=3.11.0<3.12.0
drf-nested-routers>=0.91<1.00
psycopg2>=2.8.5<2.9.0
Pillow>=7.1.2<7.2.0
flake8>=3.8.2<3.9.0
Faker>=4.1.0<4.2.0
